backend/requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-jwt-extended==4.6.0
flask-cors==4.0.0
flask-migrate==4.0.5
psycopg2-binary==2.9.9
anthropic==0.25.0
httpx==0.23.3
bcrypt==4.1.2
python-dotenv==1.0.0
alembic==1.13.1
